// src/types/RestSchedule.ts
export default interface RestSchedule {
  id?: number;
  name?: string;
  courseId?: number;
  courseName?: string;
  creditName?: number;
  testDay?: string;
  poetry?: string;
  examTime?: string;
  identificatioNumber?: number;
  examinationRoom?: string;
  userId?: number;
  userName?: string;
  userCode?: number;
}